Playing around with add_background_task.
In combination with threading, this might be the option I was looking
for. The background task starts a function that can manipulate the GUI
asynchronously. And this function can first start a thread and then wait
for it and then update the GUI. Or by removing the thread, we get a
"piecewise"-blocking function